Feat: Add test for deleteAlbum in AlbumManagerController when album is not present in DB and implementation
Test: added test for deleteAlbum method in AlbumManagerController.

Implemented logic in deleteAlbum in AlbumManagerContoller.

+++ src/main/java/com/northcoders/recordshop/controller/AlbumManagerController.java
@@ -80,10 +80,17 @@ public class AlbumManagerController {
         }
     @DeleteMapping("/{id}")
     public ResponseEntity<Album> deleteAlbum (@PathVariable("id") Long id) {
-        Album albumForDeletion = albumManagerService.getAlbumById(id);
-        albumManagerService.deleteAlbum(id);
-        return new ResponseEntity<>(albumForDeletion, HttpStatus.OK);
+        Album albumForDeletion;
+        try {
+            albumForDeletion = albumManagerService.getAlbumById(id);
+            if (albumForDeletion == null) {
+                throw new RuntimeException();
+            }
+            albumManagerService.deleteAlbum(id);
+            return new ResponseEntity<>(albumForDeletion, HttpStatus.OK);
+        } catch (RuntimeException e) {
+            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+        }
     }
-
 }
